+++ src/main/java/com/alysaa/geyserupdater/bungee/BungeeUpdater.java
@@ -1,7 +1,7 @@
 package com.alysaa.geyserupdater.bungee;
 
 import com.alysaa.geyserupdater.bungee.command.GeyserUpdateCommand;
-import com.alysaa.geyserupdater.bungee.util.BungeeJoinListener;
+import com.alysaa.geyserupdater.bungee.listeners.BungeeJoinListener;
 import com.alysaa.geyserupdater.bungee.util.Config;
 import com.alysaa.geyserupdater.bungee.util.GeyserBungeeDownload;
 import com.alysaa.geyserupdater.bungee.util.bstats.Metrics;
