+++ src/main/java/com/alysaa/geyserupdater/velocity/VelocityUpdater.java
@@ -5,7 +5,7 @@ import com.alysaa.geyserupdater.common.util.GeyserProperties;
 import com.alysaa.geyserupdater.common.util.ScriptCreator;
 import com.alysaa.geyserupdater.velocity.command.GeyserUpdateCommand;
 import com.alysaa.geyserupdater.common.util.OSUtils;
-import com.alysaa.geyserupdater.velocity.util.VelocityJoinListener;
+import com.alysaa.geyserupdater.velocity.listeners.VelocityJoinListener;
 import com.alysaa.geyserupdater.velocity.util.bstats.Metrics;
 import com.google.inject.Inject;
