+++ src/main/java/com/alysaa/geyserupdater/velocity/listeners/VelocityJoinListener.java
@@ -1,4 +1,4 @@
-package com.alysaa.geyserupdater.velocity.util;
+package com.alysaa.geyserupdater.velocity.listeners;
 
 import com.alysaa.geyserupdater.common.util.FileUtils;
 import com.velocitypowered.api.event.Subscribe;
